Contests

+++ src/main/java/dpc/auth/OAuthResponse.java
@@ -5,12 +5,12 @@ import dpc.std.StdResponse;
 /**
  * Created by jiaweizhang on 9/16/2016.
  */
-public class OAuthResponse extends StdResponse{
+public class OAuthResponse extends StdResponse {
     public String token;
-    public int userId;
+    public long userId;
     public String netId;
 
-    public OAuthResponse(int status, boolean success, String message, String token, int userId, String netId) {
+    public OAuthResponse(int status, boolean success, String message, String token, long userId, String netId) {
         this.status = status;
         this.success = success;
         this.message = message;
